+++ js/react/es5/PolyfillService.js
@@ -1,6 +1,7 @@
 "use strict";
 
 if (window.exports === undefined) { window.exports = window; }
+if (window.React === undefined && window.preact !== undefined) { var React = window.preact; }
 
 Object.defineProperty(exports, "__esModule", {
   value: true