+++ scripts/build.js
@@ -12,6 +12,9 @@
  *
  * To install dependencies download this repository and then run:
  *     npm install
+ * 
+ * To run:
+ *     node build.js
  *
  * Originally DataFormsJS React Components were built in a browser using
  * Babel Standalone rather than the full version of Babel. Babel Standalone
@@ -93,7 +96,7 @@ const isWindows = (process.platform === 'win32');
             // Add copyright
             let addCopyright;
             if (!es6) {
-                addCopyright = (file.endsWith('DataFormsJS.js') || file.endsWith('DataFormsJS.React.js'));
+                addCopyright = (file.endsWith('DataFormsJS.js') || file.endsWith('DataFormsJS.React.js') || file.endsWith('jsxLoader.js'));
             } else {
                 addCopyright = (file.endsWith('json-data.js') || file.endsWith('url-hash-router.js'));
             }
@@ -155,7 +158,7 @@ async function buildReactFiles(copyright) {
     // Read and process files one at a time
     let filesChecked = 0;
     let filesUpdated = 0;
-    const defineExports = 'if (window.exports === undefined) { window.exports = window; }';
+    const defineExports = 'if (window.exports === undefined) { window.exports = window; }\nif (window.React === undefined && window.preact !== undefined) { var React = window.preact; }';
     const allComponents = [];
     for (const component of components) {
         // Read both ES6 Class File and existing ES5 file
@@ -172,6 +175,7 @@ async function buildReactFiles(copyright) {
         // custom `require` function and allows it to use the global `React` class
         // which is already required by any custom `require` function.
         codeES6 = codeES6.replace("import React from 'react';", '');
+        codeES6 = codeES6.replace('@license', ''); // Required for all comments to be deleted
         let codeES5_New = Babel.transform(codeES6, options).code;
         if (isWindows) {
             codeES5_New = codeES5_New.replace(/\n/g, '\r\n');
@@ -187,7 +191,6 @@ async function buildReactFiles(copyright) {
 
         // Add code to array for the main [DataFormsJS.js] file
         codeES6 = codeES6.replace('export default class', 'export class');
-        codeES6 = codeES6.replace('@license', ''); // Required for all comments to be deleted
         allComponents.push(codeES6);
         filesChecked++;
     }